+++ packages/cloud-cognitive/src/components/PageHeader/index.js
@@ -6,4 +6,3 @@
 //
 
 export { PageHeader } from './PageHeader';
-export { PageActionItem } from './PageActionItem';